+++ packages/opentelemetry-instrumentation-langchain/tests/test_lcel.py
@@ -9,8 +9,9 @@ from langchain_community.utils.openai_functions import (
 from langchain_community.llms.huggingface_text_gen_inference import (
     HuggingFaceTextGenInference,
 )
-from langchain_community.chat_models import ChatOpenAI
+from langchain_community.chat_models import BedrockChat, ChatOpenAI, ChatAnthropic
 from langchain_core.pydantic_v1 import BaseModel, Field
+import boto3
 
 
 @pytest.mark.vcr
@@ -159,3 +160,72 @@ def test_custom_llm(exporter):
         == "System: You are helpful assistant\nHuman: tell me a short joke"
     )
     assert hugging_face_span.attributes["llm.completions.0.content"] == response
+
+
+@pytest.mark.vcr
+def test_anthropic(exporter):
+    prompt = ChatPromptTemplate.from_messages(
+        [("system", "You are helpful assistant"), ("user", "{input}")]
+    )
+    model = ChatAnthropic(model="claude-2")
+
+    chain = prompt | model
+    response = chain.invoke({"input": "tell me a short joke"})
+
+    spans = exporter.get_finished_spans()
+
+    assert [
+        "langchain.task.ChatPromptTemplate",
+        "langchain.task.ChatAnthropic",
+        "langchain.workflow",
+    ] == [span.name for span in spans]
+
+    anthropic_span = next(
+        span for span in spans if span.name == "langchain.task.ChatAnthropic"
+    )
+
+    assert anthropic_span.attributes["llm.request.type"] == "chat"
+    assert anthropic_span.attributes["llm.request.model"] == "claude-2"
+    assert (
+        anthropic_span.attributes["llm.prompts.0.user"] == "You are helpful assistant"
+    )
+    assert anthropic_span.attributes["llm.completions.0.content"] == response.content
+
+
+@pytest.mark.vcr
+def test_bedrock(exporter):
+    prompt = ChatPromptTemplate.from_messages(
+        [("system", "You are helpful assistant"), ("user", "{input}")]
+    )
+    model = BedrockChat(
+        model_id="anthropic.claude-3-haiku-20240307-v1:0",
+        client=boto3.client(
+            'bedrock-runtime',
+            aws_access_key_id="AKIAO0Y4W4IRWHCFPUHH",
+            aws_secret_access_key="6yXG05XtZbS0buBvBcJz2CqON1XtW8AUCQ718cgu",
+            aws_session_token="a/mock/token",
+            region_name="us-east-1"
+        ),
+    )
+
+    chain = prompt | model
+    response = chain.invoke({"input": "tell me a short joke"})
+
+    spans = exporter.get_finished_spans()
+
+    assert [
+        "langchain.task.ChatPromptTemplate",
+        "langchain.task.BedrockChat",
+        "langchain.workflow",
+    ] == [span.name for span in spans]
+
+    bedrock_span = next(
+        span for span in spans if span.name == "langchain.task.BedrockChat"
+    )
+
+    assert bedrock_span.attributes["llm.request.type"] == "chat"
+    assert bedrock_span.attributes["llm.request.model"] == "anthropic.claude-3-haiku-20240307-v1:0"
+    assert (
+        bedrock_span.attributes["llm.prompts.0.user"] == "You are helpful assistant"
+    )
+    assert bedrock_span.attributes["llm.completions.0.content"] == response.content