+++ packages/traceloop-sdk/traceloop/sdk/tracing/tracing.py
@@ -60,6 +60,7 @@ class TracerWrapper(object):
         processor: SpanProcessor = None,
         propagator: TextMapPropagator = None,
         exporter: SpanExporter = None,
+        should_enrich_metrics: bool = True,
         instruments: Optional[Set[Instruments]] = None,
     ) -> "TracerWrapper":
         if not hasattr(cls, "instance"):
@@ -118,18 +119,18 @@ class TracerWrapper(object):
 
             instrument_set = False
             if instruments is None:
-                init_instrumentations()
+                init_instrumentations(should_enrich_metrics)
                 instrument_set = True
             else:
                 for instrument in instruments:
                     if instrument == Instruments.OPENAI:
-                        if not init_openai_instrumentor():
+                        if not init_openai_instrumentor(should_enrich_metrics):
                             print(Fore.RED + "Warning: OpenAI library does not exist.")
                             print(Fore.RESET)
                         else:
                             instrument_set = True
                     elif instrument == Instruments.ANTHROPIC:
-                        if not init_anthropic_instrumentor():
+                        if not init_anthropic_instrumentor(should_enrich_metrics):
                             print(
                                 Fore.RED + "Warning: Anthropic library does not exist."
                             )
@@ -274,6 +275,10 @@ class TracerWrapper(object):
         if workflow_name is not None:
             span.set_attribute(SpanAttributes.TRACELOOP_WORKFLOW_NAME, workflow_name)
 
+        entity_name = get_value("entity_name")
+        if entity_name is not None:
+            span.set_attribute(SpanAttributes.TRACELOOP_ENTITY_NAME, entity_name)
+
         correlation_id = get_value("correlation_id")
         if correlation_id is not None:
             span.set_attribute(SpanAttributes.TRACELOOP_CORRELATION_ID, correlation_id)
@@ -353,16 +358,24 @@ class TracerWrapper(object):
         return self.__tracer_provider.get_tracer(TRACER_NAME)
 
 
-def set_correlation_id(correlation_id: str) -> None:
-    attach(set_value("correlation_id", correlation_id))
-
-
 def set_association_properties(properties: dict) -> None:
     attach(set_value("association_properties", properties))
 
 
 def set_workflow_name(workflow_name: str) -> None:
     attach(set_value("workflow_name", workflow_name))
+
+
+def set_entity_name(entity_name: str) -> None:
+    attach(set_value("entity_name", entity_name))
+
+
+def get_chained_entity_name(entity_name: str) -> str:
+    parent = get_value("entity_name")
+    if parent is None:
+        return entity_name
+    else:
+        return f"{parent}.{entity_name}"
 
 
 def set_prompt_tracing_context(
@@ -408,9 +421,9 @@ def init_tracer_provider(resource: Resource) -> TracerProvider:
     return provider
 
 
-def init_instrumentations():
-    init_openai_instrumentor()
-    init_anthropic_instrumentor()
+def init_instrumentations(should_enrich_metrics: bool):
+    init_openai_instrumentor(should_enrich_metrics)
+    init_anthropic_instrumentor(should_enrich_metrics)
     init_cohere_instrumentor()
     init_pinecone_instrumentor()
     init_qdrant_instrumentor()
@@ -429,23 +442,28 @@ def init_instrumentations():
     init_weaviate_instrumentor()
 
 
-def init_openai_instrumentor():
+def init_openai_instrumentor(should_enrich_metrics: bool):
     if importlib.util.find_spec("openai") is not None:
         Telemetry().capture("instrumentation:openai:init")
         from opentelemetry.instrumentation.openai import OpenAIInstrumentor
 
-        instrumentor = OpenAIInstrumentor(enrich_assistant=True)
+        instrumentor = OpenAIInstrumentor(
+            enrich_assistant=should_enrich_metrics,
+            enrich_token_usage=should_enrich_metrics,
+        )
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
     return True
 
 
-def init_anthropic_instrumentor():
+def init_anthropic_instrumentor(should_enrich_metrics: bool):
     if importlib.util.find_spec("anthropic") is not None:
         Telemetry().capture("instrumentation:anthropic:init")
         from opentelemetry.instrumentation.anthropic import AnthropicInstrumentor
 
-        instrumentor = AnthropicInstrumentor()
+        instrumentor = AnthropicInstrumentor(
+            enrich_token_usage=should_enrich_metrics,
+        )
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
     return True