+++ CHANGELOG.md
@@ -1,3 +1,80 @@
+## v0.16.1 (2024-04-11)
+
+### Fix
+
+- **llamaindex**: proper support for custom LLMs (#776)
+- **anthropic**: prompt attribute name (#775)
+- **langchain**: BedrockChat model name should be model_id (#763)
+
+## v0.16.0 (2024-04-10)
+
+### Feat
+
+- **instrumentation-anthropic**: Support for OpenTelemetry metrics for Anthropic (#764)
+
+### Fix
+
+- **bedrock**: support anthropic v3 (#770)
+
+## v0.15.13 (2024-04-08)
+
+### Fix
+
+- **sdk**: custom instruments missing parameters (#769)
+- **sdk**: import of removed method
+- **sdk**: removed deprecated set_context
+
+## v0.15.12 (2024-04-08)
+
+### Fix
+
+- **anthropic**: do not fail for missing methods
+- **anthropic**: Async and streaming Anthropic (#750)
+
+## v0.15.11 (2024-04-04)
+
+### Fix
+
+- **openai**: async streaming metrics (#749)
+
+## v0.15.10 (2024-04-04)
+
+### Fix
+
+- **anthropic**: token usage (#747)
+
+## v0.15.9 (2024-04-03)
+
+### Fix
+
+- **openai**: switch to init flag for token usage enrichment (#745)
+- **anthropic**: support multi-modal (#746)
+- **langchain**: instrument chat models (#741)
+
+## v0.15.8 (2024-04-03)
+
+### Fix
+
+- bump otel -> 0.45.0 (#740)
+
+## v0.15.7 (2024-04-03)
+
+### Fix
+
+- enrich spans with related entity name + support entities nesting (#713)
+
+## v0.15.6 (2024-04-02)
+
+### Fix
+
+- **sdk**: stricter dependencies for instrumentations
+
+## v0.15.5 (2024-04-02)
+
+### Fix
+
+- **openai**: missing metric for v0 instrumentation (#735)
+
 ## v0.15.4 (2024-03-31)
 
 ### Fix